requirements.txt
Telethon==1.23.0
pycoingecko==2.2.0
PyYAML==6.0
